Remove `Sites` adapter
