+++ extra/lib/plausible_web/controllers/api/external_sites_controller.ex
@@ -206,7 +206,7 @@ defmodule PlausibleWeb.Api.ExternalSitesController do
   end
 
   defp get_site(user, site_id, roles) do
-    case Plausible.Teams.Adapter.Read.Sites.get_for_user(user, site_id, roles) do
+    case Plausible.Sites.get_for_user(user, site_id, roles) do
       nil -> {:error, :site_not_found}
       site -> {:ok, site}
     end
