+++ extra/lib/plausible_web/live/funnel_settings.ex
@@ -16,7 +16,7 @@ defmodule PlausibleWeb.Live.FunnelSettings do
     socket =
       socket
       |> assign_new(:site, fn %{current_user: current_user} ->
-        Plausible.Teams.Adapter.Read.Sites.get_for_user!(current_user, domain, [
+        Plausible.Sites.get_for_user!(current_user, domain, [
           :owner,
           :admin,
           :super_admin
@@ -106,7 +106,7 @@ defmodule PlausibleWeb.Live.FunnelSettings do
 
   def handle_event("delete-funnel", %{"funnel-id" => id}, socket) do
     site =
-      Plausible.Teams.Adapter.Read.Sites.get_for_user!(
+      Plausible.Sites.get_for_user!(
         socket.assigns.current_user,
         socket.assigns.domain,
         [:owner, :admin]
