+++ extra/lib/plausible_web/live/funnel_settings/form.ex
@@ -13,7 +13,7 @@ defmodule PlausibleWeb.Live.FunnelSettings.Form do
 
   def mount(_params, %{"domain" => domain} = session, socket) do
     site =
-      Plausible.Teams.Adapter.Read.Sites.get_for_user!(socket.assigns.current_user, domain, [
+      Plausible.Sites.get_for_user!(socket.assigns.current_user, domain, [
         :owner,
         :admin,
         :super_admin
