+++ lib/plausible_web/live/imports_exports_settings.ex
@@ -15,7 +15,7 @@ defmodule PlausibleWeb.Live.ImportsExportsSettings do
     socket =
       socket
       |> assign_new(:site, fn %{current_user: current_user} ->
-        Plausible.Teams.Adapter.Read.Sites.get_for_user!(current_user, domain, [
+        Plausible.Sites.get_for_user!(current_user, domain, [
           :owner,
           :admin,
           :super_admin
