+++ lib/plausible_web/live/props_settings.ex
@@ -11,7 +11,7 @@ defmodule PlausibleWeb.Live.PropsSettings do
     socket =
       socket
       |> assign_new(:site, fn %{current_user: current_user} ->
-        Plausible.Teams.Adapter.Read.Sites.get_for_user!(current_user, domain, [
+        Plausible.Sites.get_for_user!(current_user, domain, [
           :owner,
           :admin,
           :super_admin
