+++ lib/plausible_web/live/props_settings/form.ex
@@ -18,7 +18,7 @@ defmodule PlausibleWeb.Live.PropsSettings.Form do
     socket =
       socket
       |> assign_new(:site, fn %{current_user: current_user} ->
-        Plausible.Teams.Adapter.Read.Sites.get_for_user!(current_user, domain, [
+        Plausible.Sites.get_for_user!(current_user, domain, [
           :owner,
           :admin,
           :super_admin
