+++ lib/plausible_web/live/shields/ip_addresses.ex
@@ -17,7 +17,7 @@ defmodule PlausibleWeb.Live.Shields.IPAddresses do
     socket =
       socket
       |> assign_new(:site, fn %{current_user: current_user} ->
-        Plausible.Teams.Adapter.Read.Sites.get_for_user!(current_user, domain, [
+        Plausible.Sites.get_for_user!(current_user, domain, [
           :owner,
           :admin,
           :super_admin
